+++ src/stpipe/_subproc.py
